remove BlockAll and refactor code
